requirements.txt
asgiref==3.2.10
Django==3.1.1
numpy==1.19.2
plotly==4.10.0
pytz==2020.1
retrying==1.3.3
six==1.15.0
sqlparse==0.3.1
pandas==1.2.4
pytorch==1.8.0
catboost==0.26
scikit-learn==0.24.2
